+++ controls/publish_dispatcher.rb
@@ -48,8 +48,7 @@ control 'publish-dispatcher-deny-invalidate-cache' do
     }
     # should not be able to invalidate Dispatcher cache
     # https://helpx.adobe.com/experience-manager/dispatcher/using/security-checklist.html
-    # TODO: AEM documentation states 404, but default Dispatcher config returns 403,
-    # this is due to https://github.com/shinesolutions/puppet-aem-resources/issues/29
-    it { should have_path_with_status_code('/dispatcher/invalidate.cache', 403, opts) }
+    # even though AEM documentation states 404, but default Dispatcher config returns 403, hence we're allowing both
+    it { should have_path_within_status_codes('/dispatcher/invalidate.cache', [403, 404], opts) }
   end
 end